feat: update landing page's img carousel
