+++ package.json
@@ -11,7 +11,6 @@
     "@testing-library/user-event": "^13.5.0",
     "antd": "^4.21.0",
     "axios": "^0.27.2",
-    "bootstrap": "^5.1.3",
     "craco-less": "^2.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
@@ -21,7 +20,6 @@
     "moment": "^2.29.3",
     "prettier": "^2.6.2",
     "react": "^18.1.0",
-    "react-bootstrap": "^2.4.0",
     "react-dom": "^18.1.0",
     "react-image-gallery": "^1.2.8",
     "react-router-dom": "^6.3.0",
