+++ src/assets/less/main.less
@@ -1,6 +1,11 @@
 @import './grid.less';
+@import '~antd/lib/style/themes/default.less';
 
 #root {
     height: 100%;
     width: 100%;
+}
+
+.ant-layout-content {
+    background-color: @body-background;
 }
