--- src/components/BicycleCarousel/index.less
@@ -1,9 +1,0 @@
-.carousel-img{
-img {
-    max-height: 550px;
-}
-p{
-    font-size: 20px;
-    font-weight: bold;
-}
-}
