+++ src/components/BicycleFooter/index.less
@@ -1,7 +1,11 @@
+@import '~antd/lib/style/themes/default.less';
+
+
 .footer{
     margin-bottom: 0 !important;
     padding-bottom: 0 !important;
     background-color: #fff;
+    border-top: 1px solid @border-color-base;
     .item-1 { grid-area: item-1; }
     .item-2 { grid-area: item-2; }
     .item-3 { grid-area: item-3; }
