+++ src/pages/Home/BrandList/style.less
@@ -1,7 +1,12 @@
 .brand-list {
-  margin: 20px 0;
+  margin: 60px 0;
+
   img {
     width: 100%;
     object-fit: contain;
+  }
+
+  & > h2 {
+    text-align: center;
   }
 }
