+++ src/pages/Home/index.js
@@ -1,17 +1,12 @@
 import { useState } from 'react'
-import BicycleCarousel from '../../components/BicycleCarousel'
-import BicycleTrendingList from '../../components/BicycleTrendingList'
-import './index.less'
 import { Space, Typography, Layout } from 'antd'
+import './index.less'
 
-import ebike from '../../assets/images/ebike-icon.svg'
-import mountbike from '../../assets/images/mtb-icon.svg'
-import roadbike from '../../assets/images/road-icon.svg'
-import gravbike from '../../assets/images/gravel-icon.svg'
-import kidbike from '../../assets/images/kids-icon.svg'
-
+import BicycleTrendingList from '../../components/BicycleTrendingList'
 import { BicycleFooter, HomeNavigation } from '../../components'
 import BrandList from './BrandList'
+import AdsCarousel from './AdsCarousel'
+import Banner from './Banner'
 
 const { Title, Text } = Typography
 const { Header, Content, Footer } = Layout
@@ -22,45 +17,14 @@ const Landing = () => {
       <Layout>
         <HomeNavigation />
         <Content>
-          <div className="banner">
-            <div className="banner-img">
-              <div className="banner-title">
-                <Title>WHERE THE WORLD RIDES</Title>
-              </div>
-            </div>
-            <div className="banner-bicycle-type">
-              <Title level={2}>POPULAR CATEGORIES</Title>
-              <div className="list-type-ctn">
-                <section>
-                  <img src={ebike} alt="ebike" />
-                  <Text>Electric Bikes</Text>
-                </section>
-                <section>
-                  <img src={mountbike} alt="ebike" />
-                  <Text>Mountain Bikes</Text>
-                </section>
-                <section>
-                  <img src={roadbike} alt="ebike" />
-                  <Text>Road Bikes</Text>
-                </section>
-                <section>
-                  <img src={gravbike} alt="ebike" />
-                  <Text>Gravel Bikes</Text>
-                </section>
-                <section>
-                  <img src={kidbike} alt="ebike" />
-                  <Text>Kid Bikes</Text>
-                </section>
-              </div>
+          <div className="landing-page">
+            <Banner />
+            <div className="container">
+              <AdsCarousel />
+              <BicycleTrendingList />
+              <BrandList />
             </div>
           </div>
-          <div className="carousel-slide">
-            <BicycleCarousel />
-          </div>
-         <div className="product-trending">
-            <BicycleTrendingList />
-          </div>
-          <BrandList />
         </Content>
         <BicycleFooter />
       </Layout>
